+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+Added
+  * [GH-3](https://github.com/claranet/terraform-azurerm-linux-vm/pull/3): Add `vm_os_disk` output
+
 # v7.4.0 - 2022-11-23
 
 Changed